Complete Update BookAPI

+++ complete/src/main/java/com/project4/hobookstore/service/ImageService.java
@@ -9,9 +9,11 @@ import com.project4.hobookstore.base.Constant;
 import com.project4.hobookstore.base.NotifyMessage;
 import com.project4.hobookstore.controller.BookJpaController;
 import com.project4.hobookstore.controller.ImageJpaController;
+import com.project4.hobookstore.controller.exceptions.NonexistentEntityException;
 import com.project4.hobookstore.model.Book;
 import com.project4.hobookstore.model.Image;
 import java.io.IOException;
+import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -27,47 +29,83 @@ import org.springframework.web.multipart.MultipartFile;
  */
 public class ImageService {
 
-    
-    
     private final Path root = Paths.get("uploads");
-    
+
     public ImageService() {
     }
-    
-    public List<Image> findAllImage(){
+
+    public List<Image> findAllImage() {
         ImageJpaController imgJpa = new ImageJpaController(Persistence.createEntityManagerFactory("ServerRESTfulAPIPU"));
         List<Image> listImage = new ArrayList<>();
         listImage = imgJpa.findImageEntities();
         return listImage;
     }
-    
-    public List<Image> findImagesByBid(Integer id){
+
+    public List<Image> findImagesByBid(Integer id) {
         ImageJpaController imgJpa = new ImageJpaController(Persistence.createEntityManagerFactory("ServerRESTfulAPIPU"));
         List<Image> listImage = new ArrayList<>();
         listImage = imgJpa.findImageByBId(id);
         return listImage;
     }
-    
-    public void createImage(MultipartFile file, int bid) throws IOException {
+
+    public boolean createImage(MultipartFile file, int bid) throws IOException {
         ImageJpaController imgJpa = new ImageJpaController(Persistence.createEntityManagerFactory("ServerRESTfulAPIPU"));
         BookJpaController bookJpa = new BookJpaController(Persistence.createEntityManagerFactory("ServerRESTfulAPIPU"));
         Book book = bookJpa.findBook(bid);
         String fileName = StringUtils.cleanPath(file.getOriginalFilename());
-        Image img = new Image(fileName, book);
-        imgJpa.create(img);
-      }
+        List<Image> listImage = new ArrayList<>();
+        listImage = imgJpa.findImageByBId(bid);
+        int count = 0;
+        for (int i = 0; i < listImage.size(); i++) {
+            if(listImage.get(i).getNameFile().equals(fileName)){
+                count++;
+            }
+        }
+        if(count==0){
+            Image img = new Image(fileName, book);
+            imgJpa.create(img);
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    public void updateImage(MultipartFile file, int bid) throws IOException, NonexistentEntityException, Exception {
+        ImageJpaController imgJpa = new ImageJpaController(Persistence.createEntityManagerFactory("ServerRESTfulAPIPU"));
+        BookJpaController bookJpa = new BookJpaController(Persistence.createEntityManagerFactory("ServerRESTfulAPIPU"));
+        Book book = bookJpa.findBook(bid);
+        if (book != null) {
+            String fileName = StringUtils.cleanPath(file.getOriginalFilename());
+            Image img = new Image(fileName, book);
+            imgJpa.edit(img);
+        }
+    }
     
+    public void deleteImageByBid(int bid) throws NonexistentEntityException {
+        ImageJpaController imgJpa = new ImageJpaController(Persistence.createEntityManagerFactory("ServerRESTfulAPIPU"));
+        List<Image> list = imgJpa.findImageByBId(bid);
+        for (int i = 0; i < list.size(); i++) {
+            imgJpa.destroy(list.get(i).getIid());
+        }
+    }
+
     public NotifyMessage uploadFile(MultipartFile file) {
         NotifyMessage msg = new NotifyMessage();
-        try {
-            Files.createDirectories(root);
-            Files.copy(file.getInputStream(), this.root.resolve(file.getOriginalFilename()));
-            msg.setCode(Constant.UPLOAD_IMG_SUCCESS);
-            msg.setMsg("Upload success!");
-        } catch (Exception e) {
-            throw new RuntimeException("Could not store the file. Error: " + e.getMessage());
+        String strPath = FileSystems.getDefault().getPath("").toAbsolutePath() + "\\uploads\\" + file.getOriginalFilename();
+        Path path = Paths.get(strPath);
+        if (Files.exists(path)) {
+            msg.setCode(Constant.UPLOAD_IMG_FAIL);
+            msg.setMsg("IMG exist!");
+        } else {
+            try {
+                Files.createDirectories(root);
+                Files.copy(file.getInputStream(), this.root.resolve(file.getOriginalFilename()));
+                msg.setCode(Constant.UPLOAD_IMG_SUCCESS);
+                msg.setMsg("Upload success!");
+            } catch (Exception e) {
+                throw new RuntimeException("Could not store the file. Error: " + e.getMessage());
+            }
         }
-
         return msg;
     }
 }
